requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
apache-airflow==2.10.5
apache-airflow-providers-apache-spark==4.11.0
apache-airflow-providers-common-compat==1.2.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.15.0
apache-airflow-providers-fab==1.3.0
apache-airflow-providers-ftp==3.10.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-postgres==5.12.0
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.8.2
apache-airflow-providers-ssh==3.13.1
apispec==6.6.1
argcomplete==3.5.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.1.0
blinker==1.8.2
cachelib==0.9.0
cattrs==24.1.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==3.0.3
cryptography==42.0.8
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docopt==0.6.2
docutils==0.16
email-validator==1.3.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
Faker==35.2.2
fastjsonschema==2.21.1
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.9.0
google-re2==1.1.20240702
googleapis-common-protos==1.65.0
graphviz==0.20.3
greenlet==3.1.0
grpcio==1.66.1
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.5
httpx==0.27.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
mistune==3.1.3
more-itertools==10.5.0
msgspec==0.18.6
multidict==6.1.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
numpy==1.24.4
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
pathspec==0.12.1
pendulum==2.1.2
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
prison==0.2.1
prompt_toolkit==3.0.51
propcache==0.2.0
protobuf==4.25.4
psutil==6.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
pysftp==0.2.9
pyspark==3.5.5
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
pytzdata==2020.1
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.8.1
rich-argparse==1.5.2
rpds-py==0.20.0
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
sshtunnel==0.4.0
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.15.0
tinycss2==1.2.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.5
urllib3==1.26.20
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.0.1
yarg==0.1.9
yarl==1.11.1
zipp==3.20.2
